fix: TRLST-372 fix user create
An issue was discovered in API when fixing user create in caseworker.
- Added check ensuring org is not null before attempting purge delete.
- Cherry picked a change in the release branch that fixes an API notification test. While this is unrelated, it was causing circleci build to fail

+++ trade_remedies_api/core/models.py
@@ -381,11 +381,10 @@ class User(AbstractBaseUser, PermissionsMixin, CaseSecurityMixin):
         if user_stats.get("non_draft_subs") == 0:
             Audit.objects.filter(created_by_id=self.id).delete()
             self.invitation_set.all().delete()
-            organisation = self.organisation.organisation
-            if organisation:
+            if self.organisation and self.organisation.organisation:
+                if len(self.organisation.organisation.users) == 1:
+                    self.organisation.organisation.delete()
                 self.organisation.delete()
-                if len(organisation.users) == 1:
-                    organisation.delete()
             return True
         return False
